Login Message, Eligible Vaccines, Animal list
Login API now returns: "User not found" if the username is not found in the DB.
The Vaccine Eligible list for a specific pet return all Vaccines the pet is able to receive: Vaccine per species - vaccine already received not expired
Animal List now is able to return Adoptability Status (Ready or Pending) based on the Alteration status and if the pet has received all vaccines. Also, we can use Adoptability Status for filtering.
-In development: Sorting feature based on selected parameter.