Update animalController.js
Animal Insert or update does not accept combination of either Mixed or Unknown along with other breeds. If either Mixed or Unknown is sent when adding an animal, other breeds are ignored.

To update breed from Mixed or Unknown to other, if Mixed or Unknown are selected as part of updated breeds, this update is ignored and previous mixed or unknown remain.